separate button and menu interactions
