+++ src/app/types/DEGENButtonInteraction.ts
@@ -1,0 +1,5 @@
+import { DiscordButtons } from './discord/DiscordButtons';
+export interface DEGENButtonInteraction {
+    prompt: string,
+    buttons: DiscordButtons[],
+}
