--- src/app/types/DEGENInteraction.ts
@@ -1,8 +1,0 @@
-import { MessageSelectOptionData } from 'discord.js';
-import { DiscordButtons } from './discord/DiscordButtons';
-export interface DEGENInteraction {
-    prompt: string,
-    buttons?: DiscordButtons[],
-    menuOptions?: MessageSelectOptionData[],
-    functionToCall?: string,
-}
