+++ src/app/types/DEGENMenuInteraction.ts
@@ -1,0 +1,6 @@
+import { MessageSelectOptionData } from 'discord.js';
+export interface DEGENMenuInteraction {
+    prompt: string,
+    menuOptions: MessageSelectOptionData[],
+    functionToCall: string,
+}
