+++ src/app/utils/interactions/deleteUserAddressInteraction.ts
@@ -1,12 +1,12 @@
+import { sendMenuInteraction } from '../interactionBuilders/sendMenuInteraction';
 import { DMChannel, User } from 'discord.js';
 import { DiscordUserCollection } from '../../types/discord/DiscordUserCollection';
-import { sendMessageWithInteractions } from '../sendMessageWithInteraction';
 import { ConnectedAddress } from '../../types/discord/ConnectedAddress';
 import { MessageSelectOptionData } from 'discord.js';
-import { DEGENInteraction } from '../../types/DEGENInteraction';
+import { DEGENMenuInteraction } from '../../types/DEGENMenuInteraction';
 import Log from '../Log';
 
-export const deletUserAddressInteraction = (user: User, dmChannel:DMChannel, discordUserDocument: DiscordUserCollection, connectedAddresses: ConnectedAddress[]): Promise<DEGENInteraction> => {
+export const deletUserAddressInteraction = (user: User, dmChannel:DMChannel, discordUserDocument: DiscordUserCollection, connectedAddresses: ConnectedAddress[]): Promise<DEGENMenuInteraction> => {
 	Log.debug('deleteUserAddressInteraction started.');
 
 	const menuItems: MessageSelectOptionData[] = connectedAddresses.map((address, i) => {
@@ -18,13 +18,13 @@ export const deletUserAddressInteraction = (user: User, dmChannel:DMChannel, dis
 	});
 
 
-	const dEGENInteraction: DEGENInteraction = {
+	const dEGENInteraction: DEGENMenuInteraction = {
 		prompt: 'Select the address you want to delete.',
 		menuOptions: menuItems,
 		functionToCall: 'deleteUserAddress',
 	};
 
-	return sendMessageWithInteractions(dEGENInteraction, dmChannel, user, discordUserDocument);
+	return sendMenuInteraction(dEGENInteraction, dmChannel, user, discordUserDocument);
 };
 
 // when the user hits changePOAPAddress, send them the menu interaction back
